+++ sentry/src/main/java/io/sentry/DefaultTransactionPerformanceCollector.java
@@ -56,12 +56,20 @@ public final class DefaultTransactionPerformanceCollector
     }
     if (!isStarted.getAndSet(true)) {
       synchronized (timerLock) {
-        for (ICollector collector : collectors) {
-          collector.setup();
-        }
         if (timer == null) {
           timer = new Timer(true);
         }
+        // We schedule the timer to call setup() on collectors immediately in the background.
+        timer.schedule(
+            new TimerTask() {
+              @Override
+              public void run() {
+                for (ICollector collector : collectors) {
+                  collector.setup();
+                }
+              }
+            },
+            0L);
         // We schedule the timer to start after a delay, so we let some time pass between setup()
         // and collect() calls.
         // This way ICollectors that collect average stats based on time intervals, like